Extract filesystem app repo store
